arcmenu

+++ LaGou/src/mayi/lagou/com/fragment/JobListFragment.java
@@ -42,6 +42,9 @@ public final class JobListFragment extends Fragment {
 	private String jobType = "所有职位";
 	private String city = "全国";
 	public boolean isVisible;
+	private boolean initComplate=false;
+	private boolean adapterData=false;
+	private boolean isFirstLoad=true;
 	private PullToRefreshListView mPullToRefreshListView;
 	private ListView mListView;
 	private JobItemAdapt adapter;
@@ -109,6 +112,10 @@ public final class JobListFragment extends Fragment {
 						refreshData(jobType, city, pageNum, "up", true);
 					}
 				});
+		initComplate=true;
+		if(!adapterData&&isVisible){
+			refreshData(jobType, city, pageNum, "down", true);
+		}
 	}
 
 	@Override
@@ -118,36 +125,21 @@ public final class JobListFragment extends Fragment {
 		if (isVisibleToUser) {
 			// 相当于Fragment的onResume
 			System.out.println("lagou setUserVisibleHint" + jobType);
-			if (!loadCacheData(jobType, city, pageNum, "down")) {
-				DialogUtils.showProcessDialog(getActivity(), true);
-				refreshData(jobType, city, pageNum, "down", true);
-			}
+			refreshData(jobType, city, pageNum, "down", true);
 		} else {
 			// 相当于Fragment的onPause
+			DialogUtils.hideProcessDialog();
 		}
 	}
 
 	@Override
 	public void onResume() {
 		super.onResume();
-		if (loadCacheData(jobType, city, pageNum, "down")) {
-			refreshData(jobType, city, pageNum, "down", true);
-		}
 	}
 
 	public void changeCity(String city) {
-	}
-
-	private boolean loadCacheData(String jType, String city, int pageNum,
-			final String type) {
-		final String url = LaGouApi.Host + LaGouApi.Jobs + jType + "?city="
-				+ city + "&pn=" + pageNum;
-		String responseStr = ConfigCache.getUrlCache(url, getActivity());
-		if (responseStr != null && !"".equals(responseStr)) {
-			return true;
-		} else {
-			return false;
-		}
+		this.city = city;
+		refreshData(jobType, city, pageNum, "down", true);
 	}
 
 	/**
@@ -160,38 +152,45 @@ public final class JobListFragment extends Fragment {
 		String responseStr = ConfigCache.getUrlCache(url, getActivity());
 		if (useCache && responseStr != null && !"".equals(responseStr)) {
 			List<Position> list = ParserUtil.parserPosition(responseStr);
-			if ("down".equals(type)) {
+			if ("down".equals(type)&&initComplate) {
 				allData.clear();
 				adapter.deleteAllItems();
 				adapter.addItems(list);
 				setLastUpdateTime();
 				mPullToRefreshListView.onPullDownRefreshComplete();
-			} else if ("up".equals(type)) {
+			} else if ("up".equals(type)&&initComplate) {
 				adapter.addItems(list);
 				mPullToRefreshListView.onPullUpRefreshComplete();
 			}
-			if (list != null && list.size() > 0) {
+			if (list != null && list.size() > 0&&initComplate) {
 				allData.addAll(list);
 			}
+			adapterData=true;
+			isFirstLoad=false;
 			return;
+		}
+		if(isFirstLoad){
+			DialogUtils.showProcessDialog(getActivity(), true);
 		}
 		LaGouApp.getInstance().client.get(url, new AsyncHttpResponseHandler() {
 			@Override
 			public void onSuccess(String response) {
 				List<Position> list = ParserUtil.parserPosition(response);
-				if ("down".equals(type)) {
+				if ("down".equals(type)&&initComplate) {
 					allData.clear();
 					adapter.deleteAllItems();
 					adapter.addItems(list);
 					setLastUpdateTime();
 					mPullToRefreshListView.onPullDownRefreshComplete();
-				} else if ("up".equals(type)) {
+				} else if ("up".equals(type)&&initComplate) {
 					adapter.addItems(list);
 					mPullToRefreshListView.onPullUpRefreshComplete();
 				}
-				if (list != null && list.size() > 0) {
+				if (list != null && list.size() > 0&&initComplate) {
 					allData.addAll(list);
 				}
+				adapterData=true;
+				isFirstLoad=false;
 				ConfigCache.setUrlCache(response, url);
 			}
